获取课程评价列表接口返回值优化(取消Map)

+++ sustech-course-base/sustech-course-interaction/src/main/java/edu/sustech/interaction/mapper/CourseReviewMapper.java
@@ -21,7 +21,7 @@ public interface CourseReviewMapper extends BaseMapper<CourseReview> {
      * @return 课程平均分
      */
     @Select("SELECT AVG(score) FROM course_review WHERE course_id = #{courseId} and is_delete = 0")
-    Double selectAverageScore(Integer courseId);
+    Double selectAverageScore(Long courseId);
 
     /**
      * 更新点赞数和点踩数
